[feat] 판매 금액 정산 - 모든 점주 및 OrderPayment 목록 조회, OrderPayment 상태 일괄 갱신 (#93)
* [refactor] 메뉴 가격 데이터 타입을 Long 으로 통일

* [refactor] 메뉴 가격 데이터 타입을 Long 으로 통일

* [feat] 모든 점주를 조회하는 기능

* [feat] 각 점주의 정산해야할 OrderPayment 목록을 조회하는 기능

* [feat] PointPayment -> OrderPayment 이름 변경 및 생성자 추가

* [feat] 정산해야할 OrderPayment 가 올바른지 검증하는 기능

- Vendor가 동일한지 검증
- OrderPaymentStatus 상태 검증

* [feat] OrderPayment 목록의 OrderPaymentStatus 상태를 ADJUSTMENT_SUCCESS 로 갱신하는 기능

* [style] 주문 금액 정산 기능 뼈대 코드 TODO 작성

* [feat] VendorId와 OrderPaymentStatus 기준으로 OrderPayment 목록을 조회하는 기능

* [feat] OrderPaymentStatus 를 일괄 업데이트하는 기능

- clearAutomatically 옵션 true 설정
영속성 컨텍스트에서 관리하는 엔티티와 DB 데이터의 동기화 이슈를 예방하기 위함

* [feat] OrderPaymentStatus 상태 Enum

* [feat] @Getter Lombok 추가

* [style] 에러 코드 수정

* [fix] 주문 가격 타입을 long 으로 변경하여 컴파일 오류 수정

* [fix] OrderItem 패키지 이동 변경사항 반영하여 컴파일 오류 수정

+++ src/main/java/camp/woowak/lab/menu/domain/Menu.java
@@ -35,7 +35,7 @@ public class Menu {
 	private String name;
 
 	@Column(nullable = false)
-	private Integer price;
+	private Long price;
 
 	@Column(nullable = false)
 	private Long stockCount;
@@ -44,7 +44,7 @@ public class Menu {
 	private String imageUrl;
 
 	public Menu(Store store, MenuCategory menuCategory, String name,
-				Integer price, Long stockCount, String imageUrl
+				Long price, Long stockCount, String imageUrl
 	) {
 		MenuValidator.validate(store, menuCategory, name, price, stockCount, imageUrl);
 		this.store = store;
